DeuceSSH: replace POSIX ssize_t with C99 int64_t, add extern "C" guards, cap packet size
- Replace all ssize_t with int64_t across the library, removing the
  POSIX <sys/types.h> dependency.  int64_t is standard C99 and always
  large enough for parse return values (up to 4 + UINT32_MAX).
- Remove the pointless _Static_assert (int64_t > uint32_t is true
  by definition).
- Cap max_packet_size at 64 MiB in transport_init.
- Add #ifdef __cplusplus / extern "C" guards to all 14 public headers.

+++ src/ssh/kex/dh-gex-sha256.c
@@ -51,7 +51,7 @@ serialize_bn_mpint(const BIGNUM *bn, uint8_t *buf, size_t bufsz, size_t *pos)
 	return 0;
 }
 
-static ssize_t
+static int64_t
 parse_bn_mpint(const uint8_t *buf, size_t bufsz, BIGNUM **bn)
 {
 	if (bufsz < 4)
@@ -220,7 +220,7 @@ handler(deuce_ssh_session sess)
 		{
 			size_t rpos = 1;
 			if (rpos >= payload_len) return DEUCE_SSH_ERROR_PARSE;
-			ssize_t n = parse_bn_mpint(&payload[rpos], payload_len - rpos, &p);
+			int64_t n = parse_bn_mpint(&payload[rpos], payload_len - rpos, &p);
 			if (n < 0) return (int)n;
 			rpos += n;
 			if (rpos >= payload_len) { BN_free(p); return DEUCE_SSH_ERROR_PARSE; }
@@ -269,7 +269,7 @@ handler(deuce_ssh_session sess)
 		const uint8_t *k_s = &payload[rpos];
 		rpos += ks_len;
 
-		ssize_t fn = parse_bn_mpint(&payload[rpos], payload_len - rpos, &f_bn);
+		int64_t fn = parse_bn_mpint(&payload[rpos], payload_len - rpos, &f_bn);
 		if (fn < 0) { BN_clear_free(x); res = (int)fn; goto cleanup; }
 		rpos += fn;
 
@@ -371,7 +371,7 @@ handler(deuce_ssh_session sess)
 		if (msg_type != SSH_MSG_KEX_DH_GEX_INIT) { res = DEUCE_SSH_ERROR_PARSE; goto cleanup; }
 		{
 			size_t rpos = 1;
-			ssize_t n = parse_bn_mpint(&payload[rpos], payload_len - rpos, &e_bn);
+			int64_t n = parse_bn_mpint(&payload[rpos], payload_len - rpos, &e_bn);
 			if (n < 0) { res = (int)n; goto cleanup; }
 		}
